Zetsty-Bite 26-11-2024

+++ ZetstyBite/ZetstyBite/Models/Entities/Bill.cs
@@ -7,7 +7,7 @@ public partial class Bill
 {
     public int BillId { get; set; }
 
-    public ulong BillStatus { get; set; }
+    public int BillStatus { get; set; }
 
     public int PaymentId { get; set; }
 
@@ -19,7 +19,7 @@ public partial class Bill
 
     public DateTime BillDatetime { get; set; }
 
-    public ulong BillType { get; set; }
+    public int BillType { get; set; }
 
     public virtual Account Account { get; set; } = null!;
