+++ ZetstyBite/ZetstyBite/Models/Entities/Payment.cs
@@ -3,11 +3,12 @@ using System.Collections.Generic;
 
 namespace ZetstyBite.Models.Entities;
 
+
 public partial class Payment
 {
     public int PaymentId { get; set; }
 
-    public ulong PaymentMethod { get; set; }
+    public int PaymentMethod { get; set; }
 
     public virtual ICollection<Bill> Bills { get; set; } = new List<Bill>();
 }
