+++ ZetstyBite/ZetstyBite/Models/Entities/Table.cs
@@ -3,19 +3,20 @@ using System.Collections.Generic;
 
 namespace ZetstyBite.Models.Entities;
 
+
 public partial class Table
 {
     public int TableId { get; set; }
 
     public int TableCapacity { get; set; }
 
-    public ulong TableMaintenance { get; set; }
+    public int TableMaintenance { get; set; }
 
     public int ReservationId { get; set; }
 
     public int ItemId { get; set; }
 
-    public ulong TableType { get; set; }
+    public int TableType { get; set; }
 
     public string TableStatus { get; set; } = null!;
